+++ changelog.md
@@ -1,1 +1,2 @@
-Use this file to explain what has changed in your add-on since the previous release. This will be included automatically in the release description when used with GitHub actions.
+* Supported AkelPad.
+* Added command to navigate to next/previous math formula (M/Shift+M).